Migracion de Vue 2 a Vue 3. Creacion de archivos para cada componente usando en la APP
