+++ controlEscolar/components/v-buttonBack.js
@@ -1,0 +1,8 @@
+export const vButtonBack = {
+    template:`
+    <div class="buttonBack buttonBack--HIDE">
+        <img src="https://storage.googleapis.com/cecati13/assets/arrowBack.svg" alt="Retroceder">
+        <span>REGRESAR</span>
+    </div>
+    `
+};
