+++ controlEscolar/index.html
@@ -28,7 +28,7 @@
     <footer class="footer" id="footer">        
     </footer>
     <script src="./../base.js?2022-12" type="text/javascript"></script>
-    <script src="https://unpkg.com/vue@next"></script>    
-    <script src="./controlEscolar.js?2023-02" type="text/javascript"></script>    
+    <script src="https://unpkg.com/vue@3/dist/vue.global.js"></script> 
+    <script src="./controlEscolar.js?2024-01" type="module"></script>    
 </body>
 </html>
