+++ curso-inscripcion/components/v-button.js
@@ -1,0 +1,7 @@
+export const vButton = {
+    template: `
+    <button class="button__continue">
+      Continuar
+    </button>
+    `
+  };
