+++ curso-inscripcion/components/v-buttonCancel.js
@@ -1,0 +1,5 @@
+export const vButtonCancel = {
+    template: `
+      <button> Cancelar actualización de información</button>
+    `
+  };
