+++ curso-inscripcion/components/v-buttonInscription.js
@@ -1,0 +1,7 @@
+export const vButtonInscription = {
+    template: `
+    <button class="button__inscription">
+      PRE-INSCRIBIR
+    </button>
+    `
+  };
