+++ curso-inscripcion/components/v-buttonUpdate.js
@@ -1,0 +1,7 @@
+export const vButtonUpdate = {
+    template:  `
+    <button class="button__update">
+      <slot></slot>
+    </button>
+    ` 
+};
