+++ curso-inscripcion/components/v-conexionFailBack.js
@@ -1,0 +1,8 @@
+export const vConexionFailBack = {
+    template: `
+      <div class="conexionFailBack">
+        <p>Tenemos dificultades técnicas con nuestro servidor de inscripciones, y por el momento no podemos procesar tú solicitud de inscripción.</p>
+        <p>Lamentamos los inconvenientes. Por favor intenta más tarde.</p>
+      </div>
+    `
+};
