+++ curso-inscripcion/components/v-legendUpdateData.js
@@ -1,0 +1,5 @@
+export const vLegendUpdateData = {
+    template: `
+      <p class="legendUpdateData">Todos los campos deben llenarse nuevamente para actualizar esta información.</p>
+    `
+  }
