+++ curso-inscripcion/components/v-updateContact.js
@@ -1,0 +1,18 @@
+export const vUpdateContact = {
+    methods: {
+      contactDetailCompleted(object){
+        this.$emit("updateProperties", object)
+      },
+    },
+  
+    template: `
+    <div class="">
+      <v-legendUpdateData></v-legendUpdateData>
+  
+      <v-contact
+        v-on:contactDetailCompleted="contactDetailCompleted"
+      />
+      
+    </div>
+    `
+  };
