+++ curso-inscripcion/components/v-updateSchool.js
@@ -1,0 +1,15 @@
+export const vUpdateSchool = {
+    methods: {
+      scholarshipDetailCompleted(object){
+        this.$emit("updateProperties", object)
+      },
+    },
+    
+    template: `
+    <div>
+      <v-legendUpdateData/>
+      <v-scholarship
+        v-on:scholarshipDetailCompleted="scholarshipDetailCompleted"
+      />
+    </div>  `
+  };
