+++ curso-inscripcion/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>CECATI 13. Incripción</title>    
+    <title>CECATI 13. Inscripción</title>
     <link rel="stylesheet" href="../styles.css?2022-12">
     <link rel="stylesheet" href="./curso-inscripcion.css?2023-02">
     <link rel="icon" href="../assets/favicon-96x96.png">
@@ -27,7 +27,7 @@
     <footer class="footer" id="footer"></footer>
     <script src="../base.js?2022-12" type="text/javascript"></script>
     <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11.4.34/dist/sweetalert2.all.min.js"></script>
-    <script src="https://unpkg.com/vue@next"></script>
-    <script src="./curso-inscripcion.js?2023-02-v1" type="module"></script>    
+    <script src="https://unpkg.com/vue@3/dist/vue.global.js"></script>
+    <script src="./curso-inscripcion.js?2024-01" type="module"></script>
 </body>
 </html>
